remove pro
