+++ packages/x-charts-premium/src/ChartsRadialDataProviderPremium/ChartsRadialDataProviderPremium.plugins.ts
@@ -1,11 +1,15 @@
+import {
+  RADIAL_PLUGINS,
+  type RadialPluginSignatures,
+} from '@mui/x-charts/ChartsRadialDataProvider';
 import { type PolarChartSeriesType } from '@mui/x-charts/internals';
 import {
-  RADIAL_PRO_PLUGINS,
-  type RadialProPluginSignatures,
-} from '@mui/x-charts-pro/ChartsRadialDataProviderPro';
+  useChartProExport,
+  type UseChartProExportSignature,
+} from '@mui/x-charts-pro/plugins';
 
-export const RADIAL_PREMIUM_PLUGINS = [...RADIAL_PRO_PLUGINS] as const;
+export const RADIAL_PREMIUM_PLUGINS = [...RADIAL_PLUGINS, useChartProExport] as const;
 
 export type RadialPremiumPluginSignatures<
   SeriesType extends PolarChartSeriesType = PolarChartSeriesType,
-> = [...RadialProPluginSignatures<SeriesType>];
+> = [...RadialPluginSignatures<SeriesType>, UseChartProExportSignature];
