--- packages/x-charts-pro/src/ChartsRadialDataProviderPro/index.ts
@@ -1,12 +1,0 @@
-import '../typeOverloads';
-
-export {
-  ChartsRadialDataProviderPro,
-  type ChartsRadialDataProviderProProps,
-  type ChartsRadialDataProviderProSlots,
-  type ChartsRadialDataProviderProSlotProps,
-} from './ChartsRadialDataProviderPro';
-export {
-  RADIAL_PRO_PLUGINS,
-  type RadialProPluginSignatures,
-} from './ChartsRadialDataProviderPro.plugins';
