+++ packages/x-charts-pro/src/index.ts
@@ -48,7 +48,6 @@ export type {
   ChartsContainerProSlotProps,
 } from './ChartsContainerPro';
 export * from './ChartsDataProviderPro';
-export * from './ChartsRadialDataProviderPro';
 export * from './ScatterChartPro';
 export * from './SankeyChart';
 export * from './BarChartPro';
